image_similarity/requirements.txt
flask
torch
torchvision
faiss-cpu
onnxruntime
numpy
pandas
pillow
matplotlib
tqdm